+++ components/job-automation/back/src/Infrastructure/Validation/Storage/Local/LocalStorage.php
@@ -1,9 +1,0 @@
-<?php
-
-namespace Akeneo\Platform\JobAutomation\Infrastructure\Validation\Storage\Local;
-
-use Akeneo\Platform\JobAutomation\Infrastructure\Validation\Storage\StorageConstraint;
-
-class LocalStorage extends StorageConstraint
-{
-}