+++ components/tailored-import/back/tests/Integration/Infrastructure/Validation/DataMapping/Target/Attribute/MultiSelectValidatorTest.php
@@ -143,6 +143,7 @@ final class MultiSelectValidatorTest extends AbstractValidationTest
                     'sources' => ['7fa661ce-3a6c-4b95-8441-259911b70529'],
                     'operations' => [
                         [
+                            'uuid' => 'ad4e2d5c-2830-4ba8-bf83-07f9935063d6',
                             'type' => 'split',
                             'separator' => ',',
                         ]