+++ components/tailored-import/back/tests/Specification/Application/ExecuteDataMapping/OperationApplier/SimpleSelectReplacementOperationApplierSpec.php
@@ -23,9 +23,12 @@ use PhpSpec\ObjectBehavior;
 
 class SimpleSelectReplacementOperationApplierSpec extends ObjectBehavior
 {
+    private $uuid = '00000000-0000-0000-0000-000000000000';
+
     public function it_supports_simple_select_replacement_operation(): void
     {
         $this->supports(new SimpleSelectReplacementOperation(
+            $this->uuid,
             [
                 'adidas' => ['nike', 'reebok'],
                 'foo' => ['bar', 'baz'],
@@ -36,6 +39,7 @@ class SimpleSelectReplacementOperationApplierSpec extends ObjectBehavior
     public function it_applies_simple_select_replacement_operation(): void
     {
         $operation = new SimpleSelectReplacementOperation(
+            $this->uuid,
             [
                 'adidas' => ['nike', 'reebok'],
                 'foo' => ['bar', 'baz'],
@@ -48,6 +52,7 @@ class SimpleSelectReplacementOperationApplierSpec extends ObjectBehavior
     public function it_returns_the_original_value_when_the_value_is_not_mapped(): void
     {
         $operation = new SimpleSelectReplacementOperation(
+            $this->uuid,
             [
                 'adidas' => ['nike', 'reebok'],
                 'foo' => ['bar', 'baz'],
@@ -60,6 +65,7 @@ class SimpleSelectReplacementOperationApplierSpec extends ObjectBehavior
     public function it_throws_an_exception_when_value_type_is_invalid(): void
     {
         $operation = new SimpleSelectReplacementOperation(
+            $this->uuid,
             [
                 'adidas' => ['nike', 'reebok'],
                 'foo' => ['bar', 'baz'],
@@ -71,7 +77,7 @@ class SimpleSelectReplacementOperationApplierSpec extends ObjectBehavior
 
     public function it_throws_an_exception_when_operation_type_is_invalid(): void
     {
-        $operation = new CleanHTMLTagsOperation();
+        $operation = new CleanHTMLTagsOperation($this->uuid);
         $value = new StringValue('0');
 
         $this->shouldThrow(new UnexpectedValueException($operation, SimpleSelectReplacementOperation::class, SimpleSelectReplacementOperationApplier::class))