+++ components/tailored-import/back/tests/Specification/Domain/Model/Operation/CleanHTMLTagsOperationSpec.php
@@ -19,6 +19,11 @@ use PhpSpec\ObjectBehavior;
 
 class CleanHTMLTagsOperationSpec extends ObjectBehavior
 {
+    public function let()
+    {
+        $this->beConstructedWith('00000000-0000-0000-0000-000000000000');
+    }
+
     public function it_is_initializable()
     {
         $this->shouldHaveType(CleanHTMLTagsOperation::class);
@@ -29,9 +34,15 @@ class CleanHTMLTagsOperationSpec extends ObjectBehavior
         $this->shouldBeAnInstanceOf(OperationInterface::class);
     }
 
+    public function it_returns_uuid()
+    {
+        $this->getUuid()->shouldReturn('00000000-0000-0000-0000-000000000000');
+    }
+
     public function it_normalize_operation()
     {
         $this->normalize()->shouldReturn([
+            'uuid' => '00000000-0000-0000-0000-000000000000',
             'type' => 'clean_html_tags',
         ]);
     }