+++ components/tailored-import/back/tests/Specification/Domain/Model/Operation/SplitOperationSpec.php
@@ -21,7 +21,7 @@ class SplitOperationSpec extends ObjectBehavior
 {
     public function let()
     {
-        $this->beConstructedWith(',');
+        $this->beConstructedWith('00000000-0000-0000-0000-000000000000', ',');
     }
 
     public function it_is_initializable()
@@ -34,6 +34,11 @@ class SplitOperationSpec extends ObjectBehavior
         $this->shouldBeAnInstanceOf(OperationInterface::class);
     }
 
+    public function it_returns_uuid(): void
+    {
+        $this->getUuid()->shouldReturn('00000000-0000-0000-0000-000000000000');
+    }
+
     public function it_returns_separator()
     {
         $this->getSeparator()->shouldReturn(',');
@@ -42,6 +47,7 @@ class SplitOperationSpec extends ObjectBehavior
     public function it_normalize_operation()
     {
         $this->normalize()->shouldReturn([
+            'uuid' => '00000000-0000-0000-0000-000000000000',
             'type' => 'split',
             'separator' => ',',
         ]);