+++ components/tailored-import/front/src/feature/components/DataMappingList/DataMappingList.test.tsx
@@ -1,16 +1,10 @@
 import React from 'react';
-import {screen} from '@testing-library/react';
+import {screen, act} from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import {ValidationError} from '@akeneo-pim-community/shared';
 import {Column, DataMapping} from 'feature';
-import {DataMappingList} from './DataMappingList';
 import {renderWithProviders} from 'feature/tests';
-
-const mockUuid = 'd1249682-720e-11ec-90d6-0242ac120003';
-jest.mock('akeneo-design-system', () => ({
-  ...jest.requireActual('akeneo-design-system'),
-  uuid: () => mockUuid,
-}));
+import {DataMappingList} from './DataMappingList';
 
 jest.mock('../AddDataMappingDropdown', () => ({
   AddDataMappingDropdown: ({onDataMappingAdded}: {onDataMappingAdded: (dataMapping: DataMapping) => void}) => (
@@ -37,14 +31,26 @@ jest.mock('../AddDataMappingDropdown', () => ({
 
 const dataMappings: DataMapping[] = [
   {
-    uuid: '288d85cb-3ffb-432d-a422-d2c6810113ab',
+    uuid: 'uuid-parent',
     target: {
       code: 'parent',
       type: 'property',
       action_if_not_empty: 'set',
       action_if_empty: 'skip',
     },
-    sources: ['source1', 'source3'],
+    sources: ['source1'],
+    operations: [],
+    sample_data: [],
+  },
+  {
+    uuid: 'uuid-categories',
+    target: {
+      code: 'categories',
+      type: 'property',
+      action_if_not_empty: 'set',
+      action_if_empty: 'skip',
+    },
+    sources: ['source2', 'source3'],
     operations: [],
     sample_data: [],
   },
@@ -54,7 +60,17 @@ const columns: Column[] = [
   {
     uuid: 'source1',
     index: 0,
-    label: 'Source 1',
+    label: 'parent',
+  },
+  {
+    uuid: 'source2',
+    index: 1,
+    label: 'catego 1',
+  },
+  {
+    uuid: 'source3',
+    index: 2,
+    label: 'catego 2',
   },
 ];
 
@@ -89,51 +105,7 @@ test('it can add a new data mapping', async () => {
   });
 });
 
-test('it displays the data mapping', async () => {
-  const dataMappings: DataMapping[] = [
-    {
-      uuid: 'value',
-      target: {
-        code: 'parent',
-        type: 'property',
-        action_if_not_empty: 'set',
-        action_if_empty: 'skip',
-      },
-      sources: ['source1', 'source3'],
-      operations: [],
-      sample_data: [],
-    },
-    {
-      uuid: 'another_value',
-      target: {
-        code: 'family',
-        type: 'property',
-        action_if_not_empty: 'set',
-        action_if_empty: 'skip',
-      },
-      sources: ['source2'],
-      operations: [],
-      sample_data: [],
-    },
-  ];
-  const columns: Column[] = [
-    {
-      uuid: 'source1',
-      index: 0,
-      label: 'Source 1',
-    },
-    {
-      uuid: 'source2',
-      index: 1,
-      label: 'Source 2',
-    },
-    {
-      uuid: 'source3',
-      index: 2,
-      label: 'Source 3',
-    },
-  ];
-
+test('it displays the data mappings', async () => {
   await renderWithProviders(
     <DataMappingList
       selectedDataMappingUuid={null}
@@ -146,9 +118,9 @@ test('it displays the data mapping', async () => {
     />
   );
 
-  expect(screen.getByText('pim_common.parent')).toBeInTheDocument();
+  expect(screen.getByText('pim_common.categories')).toBeInTheDocument();
   expect(
-    screen.getByText('akeneo.tailored_import.data_mapping.sources.title: Source 1 (A) Source 3 (C)')
+    screen.getByText('akeneo.tailored_import.data_mapping.sources.title: catego 1 (B) catego 2 (C)')
   ).toBeInTheDocument();
 });
 
@@ -169,7 +141,7 @@ test('it calls handler when row is selected', async () => {
 
   userEvent.click(screen.getByText('pim_common.parent'));
 
-  expect(handleDataMappingSelected).toBeCalledWith('288d85cb-3ffb-432d-a422-d2c6810113ab');
+  expect(handleDataMappingSelected).toBeCalledWith('uuid-parent');
 });
 
 test('it displays validation errors', async () => {
@@ -179,7 +151,7 @@ test('it displays validation errors', async () => {
       invalidValue: '',
       message: 'this is a data mapping validation error',
       parameters: {},
-      propertyPath: '[288d85cb-3ffb-432d-a422-d2c6810113ab]',
+      propertyPath: '[uuid-parent]',
     },
     {
       messageTemplate: 'global_error.key.name',
@@ -205,4 +177,28 @@ test('it displays validation errors', async () => {
   expect(screen.getByText('global_error.key.name')).toBeInTheDocument();
   expect(screen.queryByText('error.key.name')).not.toBeInTheDocument();
   expect(screen.getByRole('alert')).toBeInTheDocument();
+});
+
+test('it can search data mappings on column labels', async () => {
+  jest.useFakeTimers();
+
+  await renderWithProviders(
+    <DataMappingList
+      selectedDataMappingUuid={null}
+      onDataMappingSelected={jest.fn()}
+      dataMappings={dataMappings}
+      columns={columns}
+      validationErrors={[]}
+      onDataMappingAdded={jest.fn()}
+      onDataMappingRemoved={jest.fn()}
+    />
+  );
+
+  act(() => {
+    userEvent.paste(screen.getByPlaceholderText('pim_common.search'), 'cate');
+    jest.runAllTimers();
+  });
+
+  expect(screen.getByText('pim_common.categories')).toBeInTheDocument();
+  expect(screen.queryByText('pim_common.parent')).not.toBeInTheDocument();
 });