+++ deployments-ucs/timmy/timmy-create.go
@@ -1,0 +1,73 @@
+package main
+
+import (
+    "context"
+    "log"
+    "os"
+    "cloud.google.com/go/firestore"
+)
+
+func createClientFirestore(ctx context.Context, projectID string) *firestore.Client {
+    client, err := firestore.NewClient(ctx, projectID)
+    if err != nil {
+            log.Fatalf("Failed to create client: %v", err)
+    }
+
+    return client
+}
+
+func setDocument(ctx context.Context, client *firestore.Client, collection string, document string, data map[string]interface{}) {
+    result, err := client.Collection(collection).Doc(document).Set(ctx, data)
+    if err != nil {
+      log.Fatalln(err)
+    }
+    log.Print(result)
+}
+
+func deleteDocument(ctx context.Context, client *firestore.Client, collection string, document string) {
+    result, err := client.Collection(collection).Doc(document).Delete(ctx)
+    if err != nil {
+      log.Fatalln(err)
+    }
+    log.Print(result)
+}
+
+func readDocument(ctx context.Context, client *firestore.Client, collection string, document string) map[string]interface{} {
+    doc, err := client.Collection(collection).Doc(document).Get(ctx)
+    if err != nil {
+        log.Fatalf("Failed to iterate: %v", err)
+    }
+    return doc.Data()
+}
+
+func main() {
+    ctx := context.Background()
+    projectID := "akecld-blackhawk-sandbox"
+
+    // Firestore
+    clientFirestore := createClientFirestore(ctx, projectID)
+    collection := "tenant_contexts"
+    pfid := os.Args[1]
+    instance_name := os.Args[2]
+    mysql_password := os.Args[3]
+    email_password := os.Args[4]
+    deployment_version := os.Args[5]
+    data := map[string]interface{}{
+        "values": `{
+            "AKENEO_PIM_URL": "https://` + instance_name + `.bh.akeneo.ch",
+            "APP_DATABASE_HOST": "pim-mysql.` + pfid + `.svc.cluster.local",
+            "APP_INDEX_HOSTS": "elasticsearch-client.` + pfid + `.svc.cluster.local",
+            "APP_TENANT_ID": "` + pfid + `",
+            "MAILER_PASSWORD": "` + email_password + `",
+            "MAILER_URL": "smtp://smtp.mailgun.org:2525?encryption=tls&auth_mode=login&username=` + instance_name + `-akecld-bh-sandbox@mg.bh.akeneo.ch&password=` + email_password + `&sender_address=no-reply-` + pfid + `.bh.akeneo.ch",
+            "MAILER_USER": "` + instance_name + `-akecld-bh-sandbox@mg.bh.akeneo.ch",
+            "MEMCACHED_SVC": "memcached.` + pfid + `.svc.cluster.local",
+            "APP_DATABASE_PASSWORD": "` + mysql_password + `",
+            "PFID": "` + pfid + `",
+            "DEPLOYMENT_VERSION": "` + deployment_version + `",
+            "SRNT_GOOGLE_BUCKET_NAME": "` + pfid + `"
+          }`,
+    }
+    setDocument(ctx, clientFirestore, collection, pfid, data)
+    defer clientFirestore.Close()
+}