+++ src/Akeneo/ReferenceEntity/back/Domain/Model/Attribute/AttributeMaxLength.php
@@ -23,7 +23,7 @@ use Webmozart\Assert\Assert;
 class AttributeMaxLength
 {
     public const NO_LIMIT = null;
-    private const MAX_LIMIT = 65535;
+    public const MAX_LIMIT = 65535;
 
     private function __construct(
         private ?int $maxLength