+++ tests/back/Pim/WorkOrganization/Specification/TeamworkAssistant/Bundle/Doctrine/ORM/Repository/ProductRepositorySpec.php
@@ -57,10 +57,12 @@ class ProductRepositorySpec extends ObjectBehavior
         $project->getProductFilters()->willReturn([
             ['field' => 'family.code', 'operator' => 'IN', 'value' => 'guitar'],
             ['field' => 'name', 'operator' => '=', 'value' => 'Gibson Les Paul'],
+            ['field' => 'completeness', 'operator' => 'AT LEAST INCOMPLETE', 'value' => null],
         ]);
 
         $productQueryBuilder->addFilter('family.code', 'IN', 'guitar')->shouldBeCalled();
         $productQueryBuilder->addFilter('name', '=', 'Gibson Les Paul')->shouldBeCalled();
+        $productQueryBuilder->addFilter('completeness', '<', 100)->shouldBeCalled();
 
         $project->getOwner()->willReturn($user);
         $user->getGroupsIds()->willReturn([1,2]);
@@ -79,5 +81,46 @@ class ProductRepositorySpec extends ObjectBehavior
         $project->getLabel()->shouldBeCalled();
 
         $this->shouldThrow(\LogicException::class)->during('findByProject', [$project]);
+    }
+
+    public function it_converts_filters_on_completeness(
+        $productQueryBuilderFactory,
+        GetGrantedCategoryCodes $getAllGrantedCategoryCodes,
+        ProductQueryBuilderInterface $productQueryBuilder,
+        ProjectInterface $project,
+        CursorInterface $products,
+        UserInterface $user,
+        ChannelInterface $channel,
+        LocaleInterface $locale
+    ): void
+    {
+        $project->getLocale()->willReturn($locale);
+        $project->getChannel()->willReturn($channel);
+
+        $locale->getCode()->willReturn('en_US');
+        $channel->getCode()->willReturn('ecommerce');
+
+        $productQueryBuilderFactory->create([
+            'default_locale' => 'en_US',
+            'default_scope' => 'ecommerce',
+        ])->willReturn($productQueryBuilder);
+
+        $project->getProductFilters()->willReturn([
+            ['field' => 'family.code', 'operator' => 'IN', 'value' => 'guitar'],
+            ['field' => 'completeness', 'operator' => 'AT LEAST COMPLETE', 'value' => null],
+        ]);
+
+        $productQueryBuilder->addFilter('family.code', 'IN', 'guitar')->shouldBeCalled();
+        $productQueryBuilder->addFilter('completeness', '=', 100)->shouldBeCalled();
+
+        $project->getOwner()->willReturn($user);
+        $user->getGroupsIds()->willReturn([1,2]);
+        $getAllGrantedCategoryCodes->forGroupIds([1,2])->willReturn(['foo', 'bar']);
+        $productQueryBuilder->addFilter('categories', 'IN OR UNCLASSIFIED', ['foo', 'bar'], ['type_checking' => false])->shouldBeCalled();
+        $productQueryBuilder->addFilter('family', 'NOT EMPTY', null)->shouldBeCalled();
+
+        $productQueryBuilder->execute()->willReturn($products);
+
+        $this->findByProject($project)->shouldReturn($products);
     }
 }